requirements.txt
fastapi
uvicorn[standard]
celery
redis
flower
python-dotenv
sqlalchemy
psycopg2-binary
